py/pyproject.toml
[tool.poetry]
name = "imx8-coprocessor"
version = "0.1.0"
description = ""
authors = ["Isaac Parker <parrotmac@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.4"
psutil = "^5.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
